Umstrukturierung answerhandler_Json, hilfsfunktionen in neuer datei
Die hilfsmethoden für die datenbank und die json dateien liegen nun in einer neuen datei namen data_json_functions. Das spiel ist vorher getestet und alle hauptfunktionen sind spielbar
