+++ answerhandler_json.py
@@ -1,6 +1,7 @@
 import json
 import pandas as pd
 import database_SQLite as database
+import data_json_functions as djf
 from pathlib import Path
 from intentclassificator import classifyIntent, writeMessagetoTrainingData
 from phone import handleAnswer
@@ -57,16 +58,16 @@ def findAnswer(username, msg, roomId=-1):
     for elem in rooms[roomId]['triggers']:
         if elem is not None:
             elemCount += 1
-            if elem['trigName'] in msg and checkNeededStates(rooms[roomId]['triggers'][elemCount], 
-            username) and checkNeededItems(rooms[roomId]['triggers'][elemCount], username):
+            if elem['trigName'] in msg and djf.checkNeededStates(rooms[roomId]['triggers'][elemCount], 
+            username) and djf.checkNeededItems(rooms[roomId]['triggers'][elemCount], username):
                 
-                updateStates(rooms[roomId]['triggers'][elemCount], username)
+                djf.updateStates(rooms[roomId]['triggers'][elemCount], username)
                 
                 altMode = 'game'
                 altRoom = roomId
                 if elem['actions'][0] is not None:
                     for action, actionValue in zip(elem['actions'], elem['actionsValue']):
-                        altAction = doAction(action, actionValue, roomId, username)
+                        altAction = djf.doAction(action, actionValue, roomId, username)
                         if altAction[0] is not None: altRoom = altAction[0]
                         elif altAction[1] is not None: altMode = altAction[1]
                 
@@ -85,7 +86,7 @@ def findAnswer(username, msg, roomId=-1):
             elemCount += 1
             
             for name in elem['conNames']:
-                if name in msg and checkNeededStates(rooms[roomId]['connections'][elemCount], username):
+                if name in msg and djf.checkNeededStates(rooms[roomId]['connections'][elemCount], username):
                     roomId = int(elem['conRoomId'])
                     
                     return (getRoomIntroduction(roomId), getRoomName(roomId), 'game')
@@ -95,8 +96,8 @@ def findAnswer(username, msg, roomId=-1):
         for elem in rooms[roomId]['objects']:
             elemCount += 1
             
-            if elem['objName'] in msg and checkNeededStates(rooms[roomId]['objects'][elemCount], username):
-                updateStates(rooms[roomId]['objects'][elemCount], username)
+            if elem['objName'] in msg and djf.checkNeededStates(rooms[roomId]['objects'][elemCount], username):
+                djf.updateStates(rooms[roomId]['objects'][elemCount], username)
                 
                 return (elem['lookAt'], getRoomName(roomId), 'game')
 
@@ -108,7 +109,7 @@ def findAnswer(username, msg, roomId=-1):
             for elem in rooms[roomId]['items']:
                 elemCount += 1
                 
-                if elem['itemName'] in msg and checkNeededStates(rooms[roomId]['items'][elemCount], username):
+                if elem['itemName'] in msg and djf.checkNeededStates(rooms[roomId]['items'][elemCount], username):
                     
                     return (elem['lookAt'], getRoomName(roomId), 'game')
                 
@@ -118,8 +119,8 @@ def findAnswer(username, msg, roomId=-1):
             for elem in rooms[roomId]['objects']:
                 elemCount += 1
                 
-                if elem['objName'] in msg and checkNeededStates(rooms[roomId]['objects'][elemCount], username):
-                    updateStates(rooms[roomId]['objects'][elemCount], username)
+                if elem['objName'] in msg and djf.checkNeededStates(rooms[roomId]['objects'][elemCount], username):
+                    djf.updateStates(rooms[roomId]['objects'][elemCount], username)
                     
                     return (elem['lookAt'], getRoomName(roomId), 'game')
 
@@ -132,15 +133,15 @@ def findAnswer(username, msg, roomId=-1):
         for elem in rooms[roomId]['items']:
             elemCount += 1
             
-            if elem['itemName'] in msg and checkNeededStates(rooms[roomId]['items'][elemCount], username):
-                updateStates(rooms[roomId]['items'][elemCount], username)
-                add_to_inventory(elem['itemName'], roomId, username)
+            if elem['itemName'] in msg and djf.checkNeededStates(rooms[roomId]['items'][elemCount], username):
+                djf.updateStates(rooms[roomId]['items'][elemCount], username)
+                djf.add_to_inventory(elem['itemName'], roomId, username)
                 
                 return (elem['pickUp'], getRoomName(roomId), 'game')
     
     # ITEMS: NOCH NICHT FERTIG. BAUSTELLE
     elif intentID == 4:
-        return (get_inventory(roomId, username), getRoomName(roomId), 'game')
+        return (djf.get_inventory(roomId, username), getRoomName(roomId), 'game')
     
     # CURRENT ROOM: Gibt den Raumtext nochmal aus
     elif intentID == 5:
@@ -198,97 +199,6 @@ def getRoomDescription(id: int) -> str:
 
 
 """
-@author:: Cedric Nehring, Max Petendra
-@state: 10.06.20
-Check states of room.json in specified category
-Parameters
-----------
-id: the current roomid
-name: the category
-
-Returns: a list of tuples by (state, value)
-"""
-
-
-# Prüft alle angesprochenen(roomType) Zustände eines Raumes. Wenn einer nicht zutrifft wird False zurückgegeben.
-def checkNeededStates(roomElem, username: str):
-    allTrue = True
-
-    for needState, needStateValue in zip(roomElem['needStates'], roomElem['needStatesValue']):
-        if None not in (needState, needStateValue) and database.get_user_state_value(username, needState) != needStateValue:
-            allTrue = False
-
-    return allTrue
-
-def updateStates(roomElem, username: str):
-    
-    for newState, newStateValue in zip(roomElem['newStates'], roomElem['newStatesValue']):
-        if None not in (newState, newStateValue) and newStateValue:
-            database.update_user_state(username, newState, 1)
-        elif None not in (newState, newStateValue) and not newStateValue:
-            database.update_user_state(username, newState, 0)
-
-
-def checkNeededItems(roomElem, username: str):
-    allTrue = True
-
-    for needItem, needItemRoomId in zip(roomElem['needItems'], roomElem['needItemsRoomId']):
-        if None not in (needItem, needItemRoomId) and not database.does_user_item_exist(username, needItem, needItemRoomId):
-            allTrue = False
-
-    return allTrue
-
-"""
-@author:: Kristin Wünderlich
-@state: 17.06.20
-Does an action described in the room.json triggers
-Parameters
-----------
-actionName: Name of the action that should be done
-actionParam: Parameter for that action
-roomId: Id of current room
-username: Name of user
-
-Returns: (roomID, game mode)
-"""
-def doAction(actionName, actionParam, roomId, username):
-    actions = {
-        "changeMode" :1,
-        "addItem" :2,
-        "removeItem" :3,
-        "changeLocation" :4
-    }
-    actionId = actions.get(actionName, -1)
-    
-    #Action not recognized
-    if actionId == -1:
-        print("This action doesn't exist: "+actionName)
-        return (None,None)
-
-    #Change the mode
-    elif actionId == 1:
-        print("Change the mode to "+actionParam)
-        return (None, actionParam)
-    
-    #Add an item    
-    elif actionId == 2:
-        print("This item is added: "+actionParam)
-        add_to_inventory(actionParam, roomId, username)
-        return (None,None)
-    
-    #Remove an item    
-    elif actionId == 3:
-        print("This item is deleted: "+actionParam)
-        remove_from_inventory(actionParam, roomId, username)
-        return (None,None)
-
-    #Changes the room of the player
-    elif actionId == 4:
-        print("The player moves to room no.: "+ actionParam)
-        return(actionParam, None)
-
-
-"""
 @author Max Petendra
 @state 10.06.20
 handles about questions
@@ -318,39 +228,3 @@ def aboutHandler(msg: str) -> str:
     else:
         return "I didnt understand your about chatbot: question."
 
-
-# inventory stuff
-def add_to_inventory(item_name: str, room_ID, username):
-    database.insert_item(username, room_ID, item_name)
-
-
-def remove_from_inventory(item_name: str, room_ID, username):
-    database.delete_user_item(username, item_name, room_ID)
-
-
-def get_inventory(room_ID_0, username) -> str:
-    #items == list of tuples
-
-    items_db = database.get_all_user_items(username)
-
-    if items_db is None:
-        return "Your Inventory is empty"
-
-    item_str = "Your inventory contains: "
-    index = 0
-    size = len(items_db)
-    l.log_time("items_number: " + str(size))  # logging
-    for i in items_db:
-        item_name ,room_ID_1 = i
-        if index == 0:
-            item_str = item_str + item_name
-        elif index < size-1:
-            item_str = item_str + ", " + item_name
-        elif index == size-1:
-            item_str = item_str + " and " + item_name
-        else:
-            item_str = item_str + " ERROR!!! " + item_name
-            
-        index += 1
-        
-    return item_str
