+++ trainingdata.csv
@@ -109,3 +109,10 @@ ok was geht denn mit dir
 go to the market place
 cedric
 go bush
+help
+exit
+ask for price
+steal chest
+look at chest
+go to chest
+hello my friend
